Refactor profile controller with profile service
Refactored the profile controller to separate the business logic into a profile service. The controller now delegates the tasks of retrieving and updating the profile information to the service. This improves code organization and maintainability.

Created a new instance of the ProfileService class.
Implemented the getProfileInformation function to retrieve the user's profile information using the profile service.
Implemented the updateProfileInformation function to update the user's profile information using the profile service.
Utilized the responseHandler utility to send success responses with appropriate messages.
Handled errors by calling the next function with the error object.